Add files via upload
Code Created by Luis Enrique Acevedo Galicia

Date: 2019-02-03

In this notebook, the fotovoltaic energy generation is explored. Data was taken from https://www.ree.es/es/estadisticas-del-sistema-electrico-espanol/series-estadisticas/series-estadisticas-nacionales and it refers to rom the years 2007-2018, with a monthly frequency.

This data measures the real fotovoltaic generation of energy in Spain.